requirements.txt
numpy
opencv-python
tensorflow
pillow
scikit-learn
matplotlib
seaborn
